fix: Use absolute position for all Lock screens
in 9472f4eb4e200a688968d92f7743f4ed08a71577 we fixed a bug on the Lock
screen that would display the Home view and the Lock screen side by
side on some devices

To fix it we set the Lock screen view with absolute positioning

Since we found the same bug on the SetPin view. We also found that this
bug only happens when the keyboard is displayed

This increase our comprehension of the bug as the impacted view seems
to all have a `KeyboardAvoidingView` in their tree

We choose to fix the bug higher in the tree view by setting the
`LockScreenWrapper` with absolute positioning. This would allow to fix
the bug on all Lock related screens

Related commit: 9472f4eb4e200a688968d92f7743f4ed08a71577
